Changed argument types and return types of Vertex' getX,getY,setX and setY from int to double for more precise positions. Improved FieldAccess edge drawing. Wrong positioned edges came from missing filter applying to FieldAccess' edges. Removed unnecessary comments in MethodGraphLayout and SugiyamaGraph. Some types and code adjustments in other classes through different types in Vertex get and set position methods.

+++ plugins/sugiyama/src/main/java/edu/kit/student/sugiyama/graph/ISugiyamaVertex.java
@@ -18,9 +18,9 @@ public interface ISugiyamaVertex extends Vertex{
 	 */
 	public int getLayer();
 	
-	public void setX(int x);
+	public void setX(double x);
 	
-	public void setY(int y);
+	public void setY(double y);
 
 	public void setLayer(int layerNum);
